Set rows and columns in JSON config
A bit of flip-flopping here, this means that when a puzzle is saved it's
saved with it's columns and rows. That's the only change since the
equation data is already generated and saved as JSON, but in case the
sizes ever change then at least the columns and rows will be accurate
for old pre-change puzzles.